whatsapp-bot/requirements.txt
flask
python-dotenv
openai
requests
aiohttp
speechrecognition
db-sqlite3
click
pydub
backoff
